// frontend/xeno-shipment/src/pages/address-list/types.ts
export interface AddressListData {
  addressID: number
  CardCode: string
  company_name: string
  CardType: string
  full_address: string
  street1: string
  street2: string
  street3: string
  city: string
  state: string
  country: string
  postal_code: string
  contact_name: string
  contact: string
  phone: string
  email: string
  tax_id: string
  phone1: string
  website: string
  active: string
  created_userID: number | null
  created_time: string | null
  updated_userID: number | null
  updated_time: string | null
  created_user_name: string | null
  updated_user_name: string | null
  eori_number: string | null
  bind_incoterms: string | null
}